BUG: Removed force read-in of excited state for ORCA
When excited state calculations are requested in ORCA, the forces
are printed to the output in the same way as groundstate forces.
The forces of several excited states can be printed out in one SCF
step. In order to avoid a) mis-identifying excited state forces as
groundstate forces and b) arbitrarily reading in only the forces
of one excited state, io.orca is now skipping over gradients of an
excited state calculation and not reading them in.